feat: Initialize project
for the 5th time...

So far I have implemented connections to the database, migrations and a working `/addresses/:address` route, although not finished. This route is still missing the `fetchNames` query parameter.

The base API url is now `/v1`. this is to stray away from the krist-compatible URLs as this should allow for better maintainability in the future.

The history for this is a long one, but summed up it's as follows:
Started off with the Axum web framework in rust along with Sea-ORM, decided that was too much work and went to use Axum and just raw SQL, that did not want to co-operate so I rewrote it using the Actix web framework instead and writing raw SQL. Then @ElonGaties and I decided to try diesel, that was too much pain as it does not support async.

Now we are here, actix-web web framework and SeaORM. Working almost perfectly.

`/addresses/rich` conflicts with `/addresses/:address` for some fucking reason, thanks actix.